dev-requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
anyio==3.3.2
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
billiard==3.6.4.0
boto3==1.17.56
botocore==1.20.56
celery==5.0.5
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==36.0.1
Deprecated==1.2.13
dj-email-url==1.0.2
Django==3.2
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-elasticsearch-dsl==7.2.1
django-extensions==3.1.3
django-filter==2.4.0
django-log-to-telegram==2020.1.23
django-oauth-toolkit==1.6.3
django-redis==4.12.1
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
elasticsearch==7.16.0
elasticsearch-dsl==7.4.0
Faker==8.13.2
h11==0.12.0
httpcore==0.13.7
idna==3.2
jmespath==0.10.0
jwcrypto==1.0
kombu==5.0.2
multidict==5.1.0
oauthlib==3.1.1
Pillow==8.3.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.21
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.4.2
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.4
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
yarl==1.6.3
